Bug fixes, read extended disc. Added Spider.java (did'nt exist) Fixed double to float conversion in StartGamePacket.java

+++ src/java/net/blockserver/network/minecraft/StartGamePacket.java
@@ -27,9 +27,9 @@ public class StartGamePacket implements BaseDataPacket {
 		seed = level.getSeed();
 		unknown = 1;
 		gamemode = level.getGamemode();
-		spawnX = level.getSpawnpos().getX();
-		spawnY = level.getSpawnpos().getY();
-		spawnZ = level.getSpawnpos().getZ();
+		spawnX = (float) level.getSpawnpos().getX();
+		spawnY = (float) level.getSpawnpos().getY();
+		spawnZ = (float) level.getSpawnpos().getZ();
 		
 		this.buffer = ByteBuffer.allocate(29); //Not sure about this, I think its right
 		buffer.put(PID);
